refactor: bulksend 하위 패키징
