+++ src/main/java/maeilmail/bulksend/policy/PersonalSequenceChoicePolicy.java
@@ -6,7 +6,7 @@ import java.time.LocalTime;
 import java.time.Period;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import maeilmail.bulksend.ChoiceQuestionPolicy;
+import maeilmail.bulksend.sender.ChoiceQuestionPolicy;
 import maeilmail.question.QuestionCategory;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
