+++ src/main/java/maeilmail/bulksend/policy/WeekdaysPersonalSequenceChoicePolicy.java
@@ -3,7 +3,7 @@ package maeilmail.bulksend.policy;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import maeilmail.bulksend.ChoiceQuestionPolicy;
+import maeilmail.bulksend.sender.ChoiceQuestionPolicy;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
 import maeilmail.subscribe.command.domain.Subscribe;
