+++ src/main/java/maeilmail/bulksend/schedule/SendQuestionScheduler.java
@@ -1,4 +1,4 @@
-package maeilmail.bulksend;
+package maeilmail.bulksend.schedule;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -9,6 +9,10 @@ import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import maeilmail.bulksend.sender.ChoiceQuestionPolicy;
+import maeilmail.bulksend.sender.QuestionSender;
+import maeilmail.bulksend.sender.SubscribeQuestionMessage;
+import maeilmail.bulksend.view.SubscribeQuestionView;
 import maeilmail.question.QuestionCategory;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
