+++ src/main/java/maeilmail/bulksend/sender/WeeklySubscribeQuestionMessage.java
@@ -1,4 +1,4 @@
-package maeilmail.bulksend;
+package maeilmail.bulksend.sender;
 
 import java.util.List;
 import maeilmail.question.Question;
