+++ src/main/java/maeilmail/bulksend/view/SubscribeQuestionView.java
@@ -1,4 +1,4 @@
-package maeilmail.bulksend;
+package maeilmail.bulksend.view;
 
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
@@ -9,7 +9,7 @@ import org.thymeleaf.spring6.SpringTemplateEngine;
 
 @Component
 @RequiredArgsConstructor
-class SubscribeQuestionView implements MailView {
+public class SubscribeQuestionView implements MailView {
 
     private final SpringTemplateEngine templateEngine;
