+++ src/test/java/maeilmail/bulksend/sender/WeeklyQuestionMimeMessageCreatorTest.java
@@ -1,4 +1,4 @@
-package maeilmail.bulksend;
+package maeilmail.bulksend.sender;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
